+++ src/modules/m_hidemode.cpp
@@ -48,11 +48,11 @@ class Settings
 			if (modename.empty())
 				throw ModuleException("<hidemode:mode> is empty at " + tag->source.str());
 
-			unsigned int rank = tag->getUInt("rank", 0);
+			unsigned long rank = tag->getUInt("rank", 0);
 			if (!rank)
 				throw ModuleException("<hidemode:rank> must be greater than 0 at " + tag->source.str());
 
-			ServerInstance->Logs.Log(MODNAME, LOG_DEBUG, "Hiding the %s mode from users below rank %u", modename.c_str(), rank);
+			ServerInstance->Logs.Log(MODNAME, LOG_DEBUG, "Hiding the %s mode from users below rank %lu", modename.c_str(), rank);
 			newranks.emplace(modename, rank);
 		}
 		rankstosee.swap(newranks);
